requirements.txt
atlassian-python-api
robotframework
wrapt